% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\name{absoluteRiskCI}
\alias{absoluteRiskCI}
\title{Calculate upper and lower boudaries for the absolute risk (probabilities) provided by predictRisk}
\usage{
absoluteRiskCI(psp)
}
\arguments{
\item{BarrettsRiskRx}{object}
}
\value{
Data frame of per-sample low and high absolute risk with risk categories applied
}
\description{
Calculate upper and lower boudaries for the absolute risk (probabilities) provided by predictRisk
}
\author{
skillcoyne
}
