% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSWGS.R
\name{.plotSegmentedGenome}
\alias{.plotSegmentedGenome}
\title{Recommend that this not be used directly. segmentRawData generates these plots as part of the method.}
\usage{
.plotSegmentedGenome(fitted, segmented, window.depths.std,
  probes.min = NULL)
}
\description{
Recommend that this not be used directly. segmentRawData generates these plots as part of the method.
}
