% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{patientEndoscopyPlot}
\alias{patientEndoscopyPlot}
\title{Over time per endoscopy risk plot}
\usage{
patientEndoscopyPlot(brr)
}
\arguments{
\item{type}{BarrettsRiskRx object}
}
\value{
ggplot object
}
\description{
Over time per endoscopy risk plot
}
\author{
skillcoyne
}
