% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{patientRiskTilesPlot}
\alias{patientRiskTilesPlot}
\title{Per sample over time tile risk plot}
\usage{
patientRiskTilesPlot(brr, col = "Endoscopy", direction = c("fwd",
  "rev"))
}
\arguments{
\item{brr}{BarrettsRiskRx object or tibble with Risk column}

\item{col}{time column name, default 'Endoscopy'}

\item{direction}{to plot the time, fwd or rev  (forward or reverse)}
}
\value{
ggplot object
}
\description{
Per sample over time tile risk plot
}
\author{
skillcoyne
}
