% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskPrediction.R
\name{plotModelPredictions}
\alias{plotModelPredictions}
\title{Model predictions plot}
\usage{
plotModelPredictions(type = "P")
}
\arguments{
\item{type}{RR (relative risk) or P (probability, DEF)}
}
\value{
ggplot object
}
\description{
Model predictions plot
}
\author{
skillcoyne
}
