% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPlots.R
\name{plotSegmentData}
\alias{plotSegmentData}
\title{Plot genome-wide raw and segmented values for all samples in the segmentation phase.}
\usage{
plotSegmentData(brr)
}
\arguments{
\item{BarrettsRiskRx}{or SegmentedSWGS object}
}
\value{
ggplot of raw and segmented values
}
\description{
Plot genome-wide raw and segmented values for all samples in the segmentation phase.
}
\author{
skillcoyne
}
