% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\name{predictions}
\alias{predictions}
\title{Get predictions from the model}
\usage{
predictions(brr, func = c("sample", "endoscopy"))
}
\arguments{
\item{BarrettsRiskRx}{object REQ}

\item{c(sample, }{endo)}
}
\value{
predictions data frame
}
\description{
Get predictions from the model
}
\author{
skillcoyne
}
