% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\name{predictRisk}
\alias{predictRisk}
\title{Main method that should be called to use the trained model.}
\usage{
predictRisk(path = ".", raw.file.grep = "raw.*read",
  corrected.file.grep = "corr|fitted", cache.dir = NULL, verbose = T)
}
\arguments{
\item{path}{where QDNAseq raw and corrected read files are REQ}

\item{raw.file.grep}{for the filename of the raw reads file, default is 'raw.*read'}

\item{corrected.file.grep}{for the filename of the corrected reads file, default is 'corr|fitted'}
}
\value{
BarrettsRiskRx object. Additional methods that take the object provide information (predictions, rx, segmentedValues, plotSegmentData, adjustRisk, sampleResiduals)
}
\description{
Main method that should be called to use the trained model.
}
\author{
skillcoyne
}
