% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\name{predictRiskFromSegments}
\alias{predictRiskFromSegments}
\title{Get the per-sample residuals calculated from the segmentation phase.}
\usage{
predictRiskFromSegments(swgsObj, verbose = T)
}
\arguments{
\item{SegmentedSWGS}{object}
}
\value{
Predict segmented data using included model
}
\description{
Get the per-sample residuals calculated from the segmentation phase.
}
\author{
skillcoyne
}
