% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\name{relativeRiskCI}
\alias{relativeRiskCI}
\title{Calculate upper and lower boudaries for the relative risk provided by predictRisk}
\usage{
relativeRiskCI(psp)
}
\arguments{
\item{BarrettsRiskRx}{object}
}
\value{
Data frame of per-sample low and high relative risk
}
\description{
Calculate upper and lower boudaries for the relative risk provided by predictRisk
}
\author{
skillcoyne
}
