% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{riskColors}
\alias{riskColors}
\title{Get the colors used for risk classes 'low', 'moderate','high'}
\usage{
riskColors()
}
\value{
tibble
}
\description{
Get the colors used for risk classes 'low', 'moderate','high'
}
\author{
skillcoyne
}
