% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_swgs.R
\name{runQDNAseq}
\alias{runQDNAseq}
\title{Run QNDAseq on bam files using previously determined parameters and output the required raw/filtered data files}
\usage{
runQDNAseq(bam = NULL, path = NULL, outputPath = NULL,
  minMapQ = 37)
}
\arguments{
\item{outputPath}{Location to output resulting raw and fitted read files and plots}

\item{bamPath}{Location with bam files to be processed using QDNAseq}
}
\description{
Run QNDAseq on bam files using previously determined parameters and output the required raw/filtered data files
}
\author{
skillcoyne
}
