% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskPrediction.R
\name{rx}
\alias{rx}
\title{Times per sample are determined by the order of the sample as given by the demoFile, or by the order of the samples in the dataset.}
\usage{
rx(brr, demoFile = NULL)
}
\arguments{
\item{demoFile}{for the tab-delimited file that contains a per-sample entry for p53 IHC, Barrett's segment length, patient gender OPTIONAL}

\item{BarrettsRiskRx}{object REQ}
}
\value{
A table that includes recommendations per timepoint.
}
\description{
Times per sample are determined by the order of the sample as given by the demoFile, or by the order of the samples in the dataset.
}
\author{
skillcoyne
}
