% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\docType{data}
\name{rxRules}
\alias{rxRules}
\title{Times per sample are determined by the order of the sample as given by the demoFile, or by the order of the samples in the dataset.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 3 columns.}
\usage{
rxRules
}
\value{
A table describing recommendation rules
}
\description{
Times per sample are determined by the order of the sample as given by the demoFile, or by the order of the samples in the dataset.
}
\author{
skillcoyne
}
\keyword{datasets}
