% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\name{sampleNames}
\alias{sampleNames}
\title{Get the samplenames}
\usage{
sampleNames(brr, passQC = T)
}
\arguments{
\item{passQC}{logical indicating whether to return the samples that passed (T) or failed (F) QC. DEF=T}

\item{BarrettsRiskRx}{object}
}
\value{
Vector of samplenames
}
\description{
Get the samplenames
}
\author{
skillcoyne
}
