% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskPrediction.R
\name{sampleResiduals}
\alias{sampleResiduals}
\title{Get the per-sample residuals calculated from the segmentation phase.}
\usage{
sampleResiduals(brr)
}
\arguments{
\item{BarrettsRiskRx}{object}
}
\value{
Data frame of per-sample residual variance, used in QC of samples.
}
\description{
Get the per-sample residuals calculated from the segmentation phase.
}
\author{
skillcoyne
}
