% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSWGS.R
\name{scoreCX}
\alias{scoreCX}
\title{Per-sample complexity score.}
\usage{
scoreCX(df, MARGIN)
}
\arguments{
\item{df}{data frame of genome wide values}

\item{MARGIN}{columns or rows (see 'apply')}
}
\value{
cx value
}
\description{
Per-sample complexity score.
}
\author{
skillcoyne
}
