% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskPrediction.R
\name{segmentedValues}
\alias{segmentedValues}
\title{Get the segmented values for samples from the segmentation phase.}
\usage{
segmentedValues(brr, passQC = T)
}
\arguments{
\item{passQC}{logical indicating whether to return the samples that passed (T) or failed (F) QC. DEF=T}

\item{BarrettsRiskRx}{object}
}
\value{
Data frame of segmented values for samples that passed QC
}
\description{
Get the segmented values for samples from the segmentation phase.
}
\author{
skillcoyne
}
