% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSWGS.R
\name{segmentRawData}
\alias{segmentRawData}
\title{This function runs the copynumber::pcf algorithm to segment the sWGS data.
renamed from 'binSWGS'}
\usage{
segmentRawData(raw.data, fit.data,
  blacklist = read.table("inst/qDNAseq_blacklistedRegions.txt", header =
  T, sep = "\\t"), min.probes = 67, gamma2 = 250, cutoff = 0.015,
  logTransform = F, verbose = T)
}
\value{

}
\description{
This function runs the copynumber::pcf algorithm to segment the sWGS data.
renamed from 'binSWGS'
}
