% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPlots.R
\name{showModelPredictions}
\alias{showModelPredictions}
\title{Model predictions plot}
\usage{
showModelPredictions(type = "P")
}
\arguments{
\item{type}{RR (relative risk) or P (probability, DEF)}
}
\value{
ggplot object
}
\description{
Model predictions plot
}
\author{
skillcoyne
}
