% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{showPredictionCalibration}
\alias{showPredictionCalibration}
\title{Predictions risk calibration plot}
\usage{
showPredictionCalibration(df = NULL)
}
\value{
ggplot object
}
\description{
Predictions risk calibration plot
}
\author{
skillcoyne
}
