% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_swgs.R
\name{subtractArms}
\alias{subtractArms}
\title{Subtract arm-level values from smaller tile sizes. Used only if tileSegments has been run more than once with 'arms' as one tile size.}
\usage{
subtractArms(segments, arms)
}
\arguments{
\item{segments}{Matrix from tileSegments}

\item{arms}{Matrix from tileSegments (generally using 'arms' size)}
}
\value{
matrix of values
}
\description{
Subtract arm-level values from smaller tile sizes. Used only if tileSegments has been run more than once with 'arms' as one tile size.
}
\author{
skillcoyne
}
