% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSWGS.R
\name{tileSegments}
\alias{tileSegments}
\title{Presumes that the segmented data has already been filtered for num of probes, SD etc
renames 'tile.segmented.data'}
\usage{
tileSegments(data, size = 5e+06, build = "hg19", verbose = T)
}
\value{

}
\description{
Presumes that the segmented data has already been filtered for num of probes, SD etc
renames 'tile.segmented.data'
}
