% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskPrediction.R
\name{adjustRisk}
\alias{adjustRisk}
\title{Use with caution. These are based on estimates of what we think the risk might really be #' and adjusting the resulting risk based on those estimates.}
\usage{
adjustRisk(BarrettsRiskRx, offset = c("min", "mean", "max"))
}
\arguments{
\item{BarrettsRiskRx}{object}

\item{offset}{c(min,mean,max)}
}
\value{
BarrettsRiskRx object with adjusted predictions and recommentations.
}
\description{
Use with caution. These are based on estimates of what we think the risk might really be #' and adjusting the resulting risk based on those estimates.
}
\author{
skillcoyne
}
