% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{copyNumberMountainPlot}
\alias{copyNumberMountainPlot}
\title{Windowed, scaled CN values across the genome}
\usage{
copyNumberMountainPlot(brr, annotate = T, legend = T)
}
\arguments{
\item{annotate}{If true, only segments in the non-zero coefficients are highlighted (DEF=T)}

\item{type}{BarrettsRiskRx object}
}
\value{
list of ggplot objects per sample
}
\description{
Windowed, scaled CN values across the genome
}
\author{
skillcoyne
}
