% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{patientRiskTilesPlot}
\alias{patientRiskTilesPlot}
\title{Per sample over time tile risk plot}
\usage{
patientRiskTilesPlot(brr)
}
\arguments{
\item{type}{BarrettsRiskRx object}
}
\value{
ggplot object
}
\description{
Per sample over time tile risk plot
}
\author{
skillcoyne
}
