% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPlots.R
\name{plotCorrectedCoverage}
\alias{plotCorrectedCoverage}
\title{Plot genome-wide coverage from adjusted raw data}
\usage{
plotCorrectedCoverage(brr)
}
\arguments{
\item{BarrettsRiskRx}{or SegmentedSWGS object}
}
\value{
ggplot of raw and segmented values
}
\description{
Plot genome-wide coverage from adjusted raw data
}
\author{
skillcoyne
}
