% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{plotSegmentData}
\alias{plotSegmentData}
\title{Plot genome-wide raw and segmented values for all samples in the segmentation phase.}
\usage{
plotSegmentData(brr, as = c("plot", "list"))
}
\arguments{
\item{as}{Return type, plot or list of plots}

\item{BarrettsRiskRx}{or SegmentedSWGS object}
}
\value{
ggplot of raw and segmented values
}
\description{
Plot genome-wide raw and segmented values for all samples in the segmentation phase.
}
\author{
skillcoyne
}
