% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\name{predictions}
\alias{predictions}
\title{Get predictions from the model}
\usage{
predictions(brr)
}
\arguments{
\item{BarrettsRiskRx}{object REQ}
}
\value{
predictions data frame
}
\description{
Get predictions from the model
}
\author{
skillcoyne
}
