% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_swgs.R
\name{runQDNAseq}
\alias{runQDNAseq}
\title{Run QNDAseq on bam files using previously determined parameters and output the required raw/filtered data files}
\usage{
runQDNAseq(bamPath, outputPath)
}
\arguments{
\item{bamPath}{Location with bam files to be processed using QDNAseq}

\item{outputPath}{Location to output resulting raw and fitted read files and plots}
}
\description{
Run QNDAseq on bam files using previously determined parameters and output the required raw/filtered data files
}
