% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_prediction.R
\name{rx}
\alias{rx}
\title{Times per sample are determined by the order of the sample as given by the demoFile, or by the order of the samples in the dataset.
TODO This function presume that each sample is an independent timepoint. As multiple samples are typically collected per timepoint I need to allow the user to indicate that somehow and account for that in the recommendation using the maximum risk.}
\usage{
rx(brr, sample.info = NULL)
}
\arguments{
\item{sample.info}{Data frame from the 'loadSampleInformation()' function that contains a per-sample entry}

\item{BarrettsRiskRx}{object REQ}
}
\value{
A table that includes recommendations per timepoint.
}
\description{
Times per sample are determined by the order of the sample as given by the demoFile, or by the order of the samples in the dataset.
TODO This function presume that each sample is an independent timepoint. As multiple samples are typically collected per timepoint I need to allow the user to indicate that somehow and account for that in the recommendation using the maximum risk.
}
\author{
skillcoyne
}
