% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_swgs.R
\name{segmentRawData}
\alias{segmentRawData}
\title{This function runs the copynumber::pcf algorithm to segment the sWGS data.
renamed from 'binSWGS'}
\usage{
segmentRawData(info, raw.data, fit.data, blacklist = NULL,
  min.probes = 67, gamma2 = 250, norm = T, cutoff = 0.015,
  logTransform = F, cache.dir = getcachedir(), build = "hg19",
  verbose = T)
}
\arguments{
\item{raw.data}{Data frame of raw read counts (file name is also valid)}

\item{fit.data}{Data frame of fitted read values (file name is also valid)}

\item{blacklist}{qDNAseq_blacklistedRegions (defaults to file provided in package)}

\item{min.probes}{minimum number of probes per segment DEF=67  (~1Mb)}

\item{gamma2}{gamma adjustment for pcf DEF=250}

\item{norm}{normalize pcf, DEF=T}

\item{cutoff}{is the residual value cutoff for QC DEF=0.015}

\item{logTransform}{DEF=F}
}
\value{
list of objects:
'seg.vals'=segmented samples that have passed QC, 'residuals'=data frame of per-sample residuals, 'prepped.data'=adjusted raw values, 'seg.plots'=list of per-sample genome-wide plots, 'genome.coverage'=calculated genome coverage, 'failedQC'=segmented samples that have failed QC
}
\description{
This function runs the copynumber::pcf algorithm to segment the sWGS data.
renamed from 'binSWGS'
}
\author{
skillcoyne
}
